refactor: modulize home page ui: align ProviderModelSheetContent with MD3 expressive
