+++ app/src/main/kotlin/me/nanova/summaryexpressive/ui/AppNavigation.kt
@@ -15,8 +15,8 @@ import androidx.navigation3.runtime.entryProvider
 import androidx.navigation3.runtime.rememberSaveableStateHolderNavEntryDecorator
 import androidx.navigation3.ui.NavDisplay
 import me.nanova.summaryexpressive.ui.page.HistoryScreen
-import me.nanova.summaryexpressive.ui.page.HomeScreen
 import me.nanova.summaryexpressive.ui.page.OnboardingScreen
+import me.nanova.summaryexpressive.ui.page.home.HomeScreen
 import me.nanova.summaryexpressive.ui.page.settings.AdvancedSummarySetupScreen
 import me.nanova.summaryexpressive.ui.page.settings.SettingsScreen
 import me.nanova.summaryexpressive.vm.AppViewModel
