+++ app/src/main/kotlin/me/nanova/summaryexpressive/ui/page/home/LengthSelector.kt
@@ -1,0 +1,69 @@
+package me.nanova.summaryexpressive.ui.page.home
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.ButtonGroupDefaults
+import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
+import androidx.compose.material3.Text
+import androidx.compose.material3.ToggleButton
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableIntStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.semantics.role
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.tooling.preview.Preview
+import me.nanova.summaryexpressive.ui.theme.SummaryExpressiveTheme
+
+@OptIn(ExperimentalMaterial3ExpressiveApi::class)
+@Composable
+fun LengthSelector(
+    selectedIndex: Int,
+    onSelectedIndexChange: (Int) -> Unit,
+    options: List<String>,
+    enabled: Boolean,
+    modifier: Modifier = Modifier,
+) {
+    Row(
+        modifier = modifier.fillMaxWidth(),
+        horizontalArrangement = Arrangement.spacedBy(ButtonGroupDefaults.ConnectedSpaceBetween),
+    ) {
+        options.forEachIndexed { index, label ->
+            ToggleButton(
+                checked = selectedIndex == index,
+                onCheckedChange = { onSelectedIndexChange(index) },
+                enabled = enabled,
+                modifier = Modifier
+                    .weight(1f)
+                    .semantics { role = Role.RadioButton },
+                shapes = when (index) {
+                    0 -> ButtonGroupDefaults.connectedLeadingButtonShapes()
+                    options.lastIndex -> ButtonGroupDefaults.connectedTrailingButtonShapes()
+                    else -> ButtonGroupDefaults.connectedMiddleButtonShapes()
+                },
+            ) {
+                Text(label)
+            }
+        }
+    }
+}
+
+@Preview
+@Composable
+private fun LengthSelectorPreview() {
+    var selectedIndex by remember { mutableIntStateOf(0) }
+    val options = listOf("Short", "Medium", "Long")
+
+    SummaryExpressiveTheme {
+        LengthSelector(
+            selectedIndex = selectedIndex,
+            onSelectedIndexChange = { selectedIndex = it },
+            options = options,
+            enabled = true,
+        )
+    }
+}
